Wait for a reply when invalidating DDP buffers during a cancel.
This required adding support for sending a SET_TCB that requested a reply
and adding a hook back into the ddp code for replies to DDP TCBs.

While here, fix a locking issue in handle_ddp_data().  The DDP state is
locked by the socket buffer, not the inp, so move the socket buffer
lock up to protect the updates to active_id and active_count.