Always call insert_ddp_data() even if no data was placed via DDP.
This allows the DDP code to properly cleanup and dequeue any pending DDP
buffers if the chip falls out of DDP.  This might fix a panic I saw where
the software state thought a DDP buffer was queued (ddp_active_count) but
only DDP_OK was set in flags and there was queued data on the socket buffer
when the task attempted to queue a second DDP buffer.

While here, add more tracing.