Add a 'pageset' structure to rework DDP buffers and avoid holding pages.
The remaining source of CPU time in the "fast" DDP AIO path is the
calls to vm_fault_quick_hold_pages() for each AIO request.  This can be
alleviated by using a cheaper verification that a set of wired pages
matches a set of mappings via pmap_extract().  However, this requires
accessing the cached buffer much earlier (when hold_aio() is called and
a DDP index is not yet chosen).

To support this, split most of 'struct ddp_buffer' out into a new
'struct pageset' that describes a buffer used for DDP.  It holds the
backing VM pages (either held or wired) and the page pods.  A queue of
cached pagesets is held in the toep and hold_aio() tries to reuse those
before creating a new pageset.  If it can reuse it avoids the need to hold
any VM pages.  The ddp buffer is now just a simple holder for the pageset
of "active" DDP requests and the AIO request.