First pass at updating to the new AIO API.
The cancellation code is cleaner, but we now have to manage setting/
clearing cancellation functions in more places (particularly when
requeuing a request to retry), so it's mostly a wash.  This also uses
the soaio daemon pool instead of the taskqueue_thread.  Still, the
soaio daemon pool isn't a "full" taskqueue in that it suffers from the
same "task can run concurrently" bug as other taskqueues (normal sockets
use the RUNNING flag as a workaround, though they also use soref so it
has to be perfect).  We may actually need a similar scheme as right now
there is a race between release_ddp_resources() and the requeue task.

Fix some other things here:
- A job that did multiple non-DDP copies would report the length of the
  last copy instead of the full length.
- We only turned DDP off in response to a cancel.  We did not do so if
  we stopped getting AIO requests (oops!).  We may want a debounce for
  this of some sort if it turns out programs can't turn the AIO requests
  around as fast enough to keep up.  OTOH, if they aren't then that means
  we are buffering on the chip which isn't ideal either.