Allow aio_nbytes values up to SSIZE_MAX instead of just INT_MAX.
In theory this should perhaps require compat versions of all the system
calls that accept an aiocb since aio_waitcomplete() can now return a
value > INT_MAX.  This check made the aio_waitcomplete() return type
change a no-op, but with it updated it is in theory no longer one.

Given that aio_return() has been correct since 1997 and that
aio_waitcomplete() is FreeBSD-specific and not in POSIX, I'd probably
vote to just permit this particular ABI breakage.  Hopefully any old
binaries that are using aio_waitcomplete() aren't relying on
aio_read/write to fail if aio_nbytes is above INT_MAX.