Checkpoint reworking the locking and race handling for the task.
toepcb's now have a refcount, and the task references the toep, not the
socket.  We get socket and sb references from AIO jobs hung off the toep
when the task runs, not from the toep directly.  This is perhaps overly
cautious and I am perhaps handling some races that don't exist.

The requeue function is especially long and repetitive. :(