Handle a race between and indicate and queueing a DDP buffer.
An indicate might arrive while the task is copying data out of the socket
buffer.  If the task has to drop the the socket buffer lock to lock the
INP, the indicate can append data to the socket buffer and set the DDP_ON
flag while the task waits for the INP and socket buffer locks.

Handle this case by rechecking sbavail() after seeing that DDP_ON is set.
If sbavail() isn't zero, go back and try to copy again.