Add a struct resource_map object to describe a mapping.
This structure serves a couple of purposes.  First, it holds the tag and
handle used to actually access the mapping of a resource.  However, I've
named the structure so that it matches 'struct resource' and can be used
with the 'bus_read_X()', etc. helper macros.  Second, I've saved the
size of the mapping for later use in unmap_resource (which fixes the
issue I had with x86).  bus_space_unmap() requires the caller to pass in
the size and does not expect to be able to infer the size from the
tag and handle, so this seems consistent.  Finally, now that the mapping
is described by an object, we can actually choose to continue supporting
direct virtual addresses if we wish by having a 'r_vaddr' that is
equivalent to rman_get_virtual().