Disable migration of MSI-X interrupts when running under Xen.
Xen versions prior to 4.6.0 do not properly handle updates to MSI-X
table entries after the initial write.  In particular, the operation
to unmask a table entry after updating it during migration is not
propagated to the "real" table for passthrough devices causing the
interrupt to remain masked.

As a workaround, disable migration of MSI-X interrupts when running
under Xen.  A new 'machdep.disable_msix_migration' tunable can be
set to 0 to always permit migration or 1 to always deny migration.