Update the manpage for the event mask.
This includes adding a section talking about optional events and
documenting each event flag.  Other places are also updated to use
the event flags rather than PT_FOLLOW_FORK and PT_LWP_EVENTS.  The
description of PT_FOLLOW_FORK and PT_LWP_EVENTS are reworked to
be in terms of modifying the mask.  In addition, the description of
PT_SYSCALL, PT_TO_SCE, and PT_TO_SCX are reworked to both indicate the
effect on the event mask but also to mention the PT_CONTINUE effect of
these requests (which was not clear to me from the previous descriptions).