Adjust t4_port_init() to work with VF devices.
Specifically, the FW_PORT_CMD may or may not work for a VF (the PF
driver can choose whether or not to permit access to this command).
Don't attempt to fetch port information on a VF if permission is
denied by the PF.  In addition, move the command down to the point
where its results are used to make the code flow less convoluted in
the VF case.