Import SSE2, AVX, and ERMS variants of memcpy, memmove, and bcopy.
For now export these symbols in the private namespace to permit testing.
Eventually we should have an ifunc that chooses the best method for the
current CPU.

One optimization we do not do (it would remove a branch) would be to
remove the backwards code from memcpy and assume memcpy never overlaps.
This would be inconsistent with our other memcpy implemtentations.