Make device_quiet() an attachment property.
In particular, reset the DF_QUIET flag when detaching from a device so
that a driver that marks a device quiet doesn't dictate policy for a
different driver that may claim the device in the future.

Remove explicit device_verbose() from the t4iov driver detach routine
now that this case is handled generically.