Test for lengths of 0 and fix bugs with lengths of 0.
The unaligned SSE2 and AVX cases would treat 0 as infinity.  Rework
them to do the remainder copy at the end and skip to it right away.
This simplifies the flow a bit and handles lengths of 0 correctly.

Also, AVX was following the backwards 128 byte copy with only a 64
byte update in lengths.