Checkpoint HMAC.
Given that swcrypto doesn't bother with plain SHA1, just punt on that
(it was going to be a one-off anyway) and start on the various HMAC
digests in hash-only mode.  So far, keying is handled and I've started
on populating the request for HMAC, but much more to do here.