Construct the firmware work request header.
Also, fix a bug in the location of the nested cpl_rx_phys_dsgl and
add a dummy tx_channel_id (still set to 0) to mark the places that
will need to use it if it is ever non-zero.

In theory this is enough to send down a hmac request now.