KSTACK_LARGE_PAGE bug fix.
Make the odd page of the TLB entry the kernel stack be the kernel
stack page and the even page the kernel stack guard page. Previously
these were swapped and the odd page wasn't being mapped in the virtual
address and being used by other 4K pages in the kernel.