Change input SGLs to honor crd restrictions.
We can't just assume that the sglist length for the crp equals what is
needed for the crd.  Instead, we need to cap the input SGL to honor
crd_skip and crd_len.  However, use this opportunity to make including
crd_skip optional and use it for cases where the IV is already in the
payload.  Howver, for the encrypt case where the IV isn't present, don't
use crypto_copyback, but instead store the IV in the immediate data by
skipping crd_len.