Fix some GCM IV confusion.
CRYPTO_AES_NIST_GMAC is not GCM with a 12-byte IV (and GCM_16 with a 16-byte
IV).  GMAC appears to just be a hash that /dev/crypto doesn't support and
nothing but cryptosoft implements, so just drop it.  Always use a 12-byte
IV with GCM via CRYPTO_AES_NIST_GCM_16.