Handle OCF's requirements for authenc decryption requests.
Set the 'hash_size' in the work request to request a copy of the calculated
hash in the reply.  If we can a "MAC" error for a decryption request, ignore
the error and copy the MAC out of the reply CPL into the buffer.  If we
didn't get a MAC error, then the input hash must have been correct so just
leave it in place.