Fix AUTHENC for IPSec packets.
IPSec packets are described by an mbuf that contains a header that is not
authenticated.  The T6 crypto engine just copies over any data not
described by one of the AAD, auth, or cipher regions.  Adjust the request
buffer sent to the card to exclude the header.  This requires additional
fixups when computing the various ranges.