Add a new GDB_LIBEXEC option to install gdb and kgdb to /usr/libexec.
When this option is enabled, only gdb and kgdb are installed to /usr/libexec
for use by crashinfo(8).  Other bits of GDB such as gdbserver and gdbtui
are not installed.  For this option to be effective, GDB must be enabled.

Rework r317094 to re-enable GDB on all platforms but enable GDB_LIBEXEC
on platforms for which the GDB in ports is a superset of functionality.