First cut at a test utility for the OpenSSL crypto engine.
sslcheck is similar in function to cryptocheck.  However, instead of
operating on /dev/crypto directly, it performs OpenSSL operations with
both a NULL engine (for the software control) and the "cryptodev"
engine.

It uses the recently added "DEVICE" and "CRID" control knobs to
implement the -d option to force the engine to use specific crypto
devices.