Fix an issue with GCM IV generation.
EVP_CTRL_GCM_IV_GEN needs to apply the 64-byte increment when generating an
IV for the next request, not affecting the IV of the current request.  To
fix this, add a new gcm_iv[] array that is used for IV generation.  The
SET_IV_FIXED control operation now populates this new array and does not
alter ctx->iv[].  The GCM_IV_GEN and GCM_SET_IV_INV control operations
use the contents of the gcm_iv[] to populate ctx->iv[] along with setting
'gcm_iv_set'.  The counter increment for GCM_IV_GEN is done on the template
in gcm_iv[] after it has been copied to ctx->iv[].