Add a new KTR_STRUCT_ARRAY which dumps an array of structures.
The structure name in the record payload is preceded by a size_t
containing the size of the individual structures.  Use this to
replace the previous code that dumped the kevent arrays dumped for
kevent().  kdump is now able to decode the kevent structures rather
than dumping their contents via a hexdump.

One change from before is that the 'changes' and 'events' arrays are
not marked with separate 'read' and 'write' annotations in kdump output.
Instead, the first array is the 'changes' array, and the second array
(only present if kevent doesn't fail with an error) is the 'events'
array.  For kevent(), empty arrays are denoted by an entry with an array
containing zero entries rather than no record.