Initial import of kernel TLS from Netflix
This is the kernel side of the Netflix kernel TLS system.
Note that this is based on the original work of scottl, and
rrs, with contributions from jtl and glebius.  It was re-written
by gallatin to use the same not-ready mbuf mechanisms for
queuing as sendfile.

It requires a backend crypto module to perform the actual
encryption.

It requres hooks in userspace to enable TLS, which will be
coming in a separate commit.

Sponsored by: Netflix

Add an SB_TLS_IFNET flag for sb_tls_flags.

When this flag is set, encryption is performed in the ifnet layer rather
than above the protocol layer.  Don't call sbtls_encrypt() when this flag
is set, don't mark mbufs as M_NOTREADY, and don't pass PRUS_NOTREADY to
pru_send() when this flag is set on a socket buffer.

This required changing sbtls_frame() to accept the socket instead of the
tls structure so that it can see the state of SB_TLS_IFNET flag.

Add a hack to disable the m_copydata path in TCP output for ifnet TLS.

By copying the raw data into a new mbuf, the TLS session info is lost
and the data is sent unencrypted.

In some cases, the first TLS record (FINISH) can be appended to the
socket buffer while other data is still waiting to be sent.  To handle
this case, always force out any non-TLS data first before sending TLS
records on an ifnet TLS connection.