Flesh out 'wait' support.
If the port number passed to -G is prefixed with 'w', enable 'wait' mode
which pauses vcpu 0 at the first instruction until a debugger attaches
and then resumes it.

The old implementation attempted to wait for the GDB connection during
startup, but that will then see the connection before the vcpu is ready.
Instead, add a check in gdb_cpu_add() to see if vcpus are suspended when
a new cpu is added and if so suspend the new vcpu.  (This should also fix
the case of single stepping a vcpu during AP startup by pausing the AP
as soon as it starts.)  For the wait case, mark vcpu 0 as suspended
(even though it isn't active yet) and set first_stop to not report the
initial stop (since the debugger won't be attached yet).  Now vcpu 0's
thread will start and immediately report a debug vmexit before executing
and doing the normal suspend in the vcpu thread.  The initial debugger
connection occurs via the kevent handling in the mevent thread.