Don't discard AAD and IV output data for AEAD requests.
The T6 can hang when processing certain AEAD requests if the request sets
a flag asking the crypto engine to discard the input IV and AAD rather than
copying them into the output buffer.  The existing driver always discards
the IV and AAD as we do not need it.  As a workaround, allocate a single
"dummy" buffer when the ccr driver attaches and change all AEAD requests to
write the IV and AAD to this scratch buffer.  The contents of the scratch
buffer are never used (similar to "bogus_page"), and it is ok for multiple
in-flight requests to share this dummy buffer.

Submitted by:	Harsh Jain @ Chelsio