Require the SHF_ALLOC flag for sections loaded from kernel object modules.
ELF object files can contain sections which are not supposed to be loaded
into memory (e.g. .comment).  Normally the static linker uses these flags
to decide which sections are allocated to loadable program segments in
ELF binaries and shared objects (including kernels and DSO-style kernel
modules on architectures other than amd64).

The kernel and loader were not checking the SHF_ALLOC flag but loading
any sections with certain types meaning that when kernel modules for
amd64 were loaded, extra memory was used to hold unneeded data such as
the contents of .comment and a few other sections.  With GNU BFD linker
this was mostly harmless (aside from wasting a small amount of RAM)
since all of these unallocated sections were placed at the end of the
file.  LLD does not sort the sections, so unallocated sections were
mixed in with allocated sections.  Since the kernel was consistent about
applying symbol relocations, having this extra data mixed in between
allocated sections did not result in any runtime failures.  However, it
does confuse other tools that read the ELF headers of kernel modules to
determine how a module's memory is laid out (that is, debuggers).  In
particular, kgdb was computing different offsets for symbols and not
displaying the correct data to the user for any sections loaded after
an unallocated section.  mdb happened to be immune from this because I
copied the same bug from link_elf_obj.c into mdb but will be fixed.
A kernel-aware lldb would also probably be confused.