Switch the MIPS N32/N64 architectures to assuming hard-float.
This retires the mips64*hf architectures and changes the default for the
mipsn32 and mips64* architectures to hard-float instead of soft-float.
The existing O32 architectures are left as-is.  Architectures supporting
soft-float for N32 and N64 ABIs are not provided by default (though they
could be easily added if need be).

All of the N32/N64 hardware that FreeBSD runs on should include an FPU,
so soft-float variants of those ABIs are not needed.

It is not clear if we should provide a libsoft fallback for old
N32 and N64 binaries.