Initial import of kernel TLS from Netflix
This is the kernel side of the Netflix kernel TLS system.
Note that this is based on the original work of scottl, and
rrs, with contributions from jtl and glebius.  It was re-written
by gallatin to use the same not-ready mbuf mechanisms for
queuing as sendfile.

It requires a backend crypto module to perform the actual
encryption.

It requres hooks in userspace to enable TLS, which will be
coming in a separate commit.

Sponsored by: Netflix