Some WIP hacking on a solution.
Add a flag to mark a KSI as seen by ptrace and avoid reporting it a
second time if so.  I was going to add a variant of ptracestop that
the sending thread would block in, but that would turn kill into a
blocking system call which is really bad if the debugger is the one
calling kill.