Simplify the socket reference handling for the TLS case.
This does add an additional soref/rele for the case of doing TLS on
buffers not ready due to sendfile I/O but is easier to reason about in
terms of correctness.  Now, sbtls_frame() sets pg_so (needed anyway for
the SB_TLS_IFNET case) instead of sbtls_enqueue(), and soref() is called
then to obtain a reference.  Callers of sbtls_enqueue() do not need to
explicitly obtain a reference as that is now done in sbtls_frame().  The
reference is held until it is released either by sbtls_encrypt() or
mb_free_ext() (the latter is true for the SB_TLS_IFNET case).