Finish initial conversion of p_xthread to a list of threads.
Threads now queue themselves on p_xthreads in ptracestop() and are removed
by continue requests to ptrace().  Normally threads are added to the end,
except for the FSTP case which adds the thread to the front.  Suspended
threads are removed from the queue for PT_SUSPEND and readded to the tail
of the queue for PT_RESUME.