Fix a TLS reference counting issue
The TLS soref() was changed from a per-enqueue reference count
that used one reference across multiple TLS records to
a per-frame reference count that uses one reference per
TLS record. This introduced a bug where sblts_encrypt wasn't
freeing all the socket references (just the first in the group).

To avoid a detour through a delayed free, we must
ensure that sblts_encrypt() is properly dropping
the reference for each record.  This is slightly challenging,
because we want to hold on to the references until pru_ready()
is called so that the so pointer is valid.   So rather than
immediately dropping the reference in the main loop, we count
references and drop them after the loop, and after calling
pru_ready.