Fix warnings from soref cleanups.
CURVNET_SET() declares variables used in CURVNET_RESTORE(), so they need
to be used at the same relative scope to work correctly.  The previous
patch was using the uninitialized variables declared in the CURVNET_SET()
before pru_ready() in the drop case.  Instead, initialize error to 0 and
break out of the loop if we get an encryption error.  Then use the existing
CURVNET_SET, but only call pru_ready() if error is 0.