Send an empty TCP packet with any TCP options before each TLS "packet".
If the TCP header contains any options other than NOPs, send a single
TCP packet containing only the options and no payload before sending the
segments from the TLS record(s).  For now I'm using this for timestamps,
but this is approach is probably not well suited to timestamps.  It probably
will work ok for SACK though since SACK options are not sent very often.