Correct masks for global register settings for KERN_TLS.
These were mostly harmless as the shift count for IPTTL and WRTHRTHRESH
were both 0.  RDTHRESHOLD was 1 bit wider than WRTHRTHRESH, but that
extra bit is the F_WRTHRTHRESHEN flag we are already setting.  The upshot
of all this is that we weren't changing the read threshold when disabling
TCB caching for debugging.