Rework timestamp support.
Instead of using TCB fields, rewrite the value of the timer used to set
the timestamps periodically.  Track the last timestamp value written to
the timer and write an updated value if a host-provided timestamp is
different.  The 32-bit host timestamp is written into bits 31:62 since
the highest possible window of the timer is bits 31:62.  The lower 30
bits are set to 0.  The timer will increment after being written, but
it will take 3 or 4 seconds before it accumulates enough ticks to modify
bit 31's value.  However, TCP timestamps are updated much more frequently
than every 3 seconds, so in practice the low 30 bits will be reset to 0
long before they will wrap.

If there is an idle gap in the traffic, the timer will run free, but the
next attempt to transmit a TLS record will reset the timer to the current
host timestamp value before the TLS record is sent.

Since this approach does generate timestamps matching the host-provided
timestamps, stop sending empty packets if the only TCP option is the
timestamp option.