Fix an mbuf leak.
A single TLS TSO request can span multiple work requests.  Since the NIC
can return credits for individual work requests, each work request must
have a valid software descriptor holding the credits used by each individual
work request.  However, the mbuf chain for the request should not freed
until all off the requests have completed.  This is implemented by only
storing the mbuf pointer in the last software descriptor used.

Previously this was implemented by checking if a given TLS record's mbuf
was the final mbuf in the chain and storing the mbuf pointer in the
software transmit descriptor if so.  However, if the final mbuf in a chain
was for a very short TLS record such that we did not actually transmit
anything (the length was less than the TLS header), the last mbuf would
not actually use a software transmit descriptor.  As a result, no
descriptor would reference the mbuf chain and it would not be freed when
the transmit was complete.

To fix, keep track of the pointer to the mbuf "slot" in the last descriptor
used by each invidiual TLS work request.  The very short TLS record case
doesn't update this pointer whereas all TLS records that generate a work
request update it to point to their associated software descriptor.  Once
a single TSO request has finished processing all of the TLS records, it
saves the mbuf pointer in the current value of this pointer which should
point to the most recently used descriptor.