Drop PUSH and FIN from options-only packets.
The PUSH flag will be set on the final segment of the final TLS record
when all of the data has been received.  FIN needs to be deferred to the
end.  Right now we don't really handle FIN on a TLS record but this at
least avoids possibly sending it early.