Rework how the last work request is detected and fix a PUSH issue.
sbtls_write_tls_wr() needs to know which work request is the final both to
set the mbuf pointer in the tx descriptor (see earlier change to fix an
mbuf leak) and to set TH_PUSH correctly.  As such, add logic to detect this
in sbtls_write_tls_wr() and push the setting of the mbuf pointer back down
into this function.