Be more careful with writing immediate data into the work request.
Especially with combo WRs, we shouldn't assume that anything other than a
16-byte aligned chunk might not wrap around the end of the ring.  To fix
this, use copy_to_txd() for inline key contexts and the immediate data
including the TLS header and IV.  Convert 'iv' to a char array on the
stack.  Reuse the scratch buffer used for SET_TCB_FIELD for the TLS header.