Slightly simplify the work request generation via a scratch buffer.
The mix of txq_advance vs copy_to_txd was getting a bit crazy and it
seems fragile for future maintenance.  Instead, add a scratch buffer
to each txq that can hold a single work request.  If we are near the
end of the tx queue such that a work request might wrap, write the
work request into the scratch buffer and then use a single copy_to_txd()
to safely copy the scratch buffer into the ring while handling wrap-around.
If we are far enough from the end of the ring that we won't wrap, write
the work request directly into the ring, but without extra complications
to handle possible wrap-around.