Handle requests to transmit just the header of a TLS record.
The stack may send a request to transmit only the header (or a portion)
of a TLS record.  Previously we just ignored these requests hoping the
stack would eventually ask to send more.  Now we will handle these
requests by constructing a TCP packet and sending it as a simple tunnelled
packet.

These packets should be small and the TLS header contents need to be
generated (along with possibly updating the TCP sequence number in the
TCP header), so always use a work request with the packet as immediate
payload.

Add a new 'kern_tls_header' stat to count these packets.

While here, fix the 'options-only' TCP packet path to actually clear
PUSH and FIN from the options-only packet.  The comment was there and
all the work to make a modifiable copy of the TCP header, but not the
line to clear the bits.

This also simplifies the check for the "last" TLS record in a chain
as that no longer has to explicitly ignore header-only records.