Handle an edge case for "middle" records.
Suppose a TLS record request has an offset that starts in the tag but
does not fully cover the tag.  In that case, we treat the request as a
short record and drop the tag.  However, when computing the "offset" of
the AES-CTR region we could end up with an offset that exceeded the
AES region.  Cap the offset into the TLS record to ensure it stays within
the record.  In theory for such a request we really should send no data
at all (we can't send the partial tag the host has requested and the host
doesn't want to send any of the TLS record payload), but this patch
instead sends out the last AES block of the payload to at least provoke
an ACK from the remote end even though we are sending data that has already
been sent (possibly even already ACKed).