More for the expanded commit note.
This change reverts r189404 which worked around a buggy BIOS which enumerated
an I/O APIC twice (using the same memory mapped address for both entries but
using an IRQ base of 256 for one entry and a valid IRQ base for the second
entry).  Making the "base" of MSI IRQ values dynamic avoids the panic that
r189404 worked around, and there may now be valid I/O APICs with an IRQ base
above 256 which this workaround would incorrectly skip.

If in the future the issue reported in PR 130483 reoccurs, we will have to
add a pass over the I/O APIC entries in the MADT to detect duplicates using
the memory mapped address and use some strategy to choose the "correct"
one.