First cut at splitting KERN_TLS support out to a separate module.
This module depends on t4_tom.ko for now, in particular for routines in
tom/t4_l2t_tom.c.  However, I think it won't require tom to actually be
enabled on the interface.  For now, the existing KERN_TLS code remains
in t4_tls.c but is effectively #if 0'd.