Use TLS L2T entries instead of the TOM L2T entry allocator.
The TOM allocator wasn't really suitable for TLS sessions since it does its
own routing resolution and we were just relying on that matching up with
the destination MAC address in TLS packets.  Now we always use an L2T entry
whose MAC addresses matches the stack-supplied destination MAC address.
In addition, the TOM code had this async behavior with failure handling
that wasn't appropriate for TLS sessions.

Now, we no longer allocate an L2T entry during connection setup, but
allocate them when sending the first packet, or if the destination MAC
address of a packet doesn't match the current L2T entry (so it will handle
the MAC changing during a connection).  This requires potentially one more
TCB field update for a TLS work request.  It also potentially requires an
additional work request to populate the the L2T entry if a new entry is
allocated (as opposed to gaining a reference on an existing entry).

This also removes the last dependency on t4_tom.ko.